(refactor) seperate db connection and helpers into database/
