+++ database/index.js
@@ -1,0 +1,27 @@
+const mongoose = require('mongoose');
+const { MONGO_URI } = process.env;
+
+mongoose.connect(MONGO_URI, { useMongoClient: true })
+
+const db = mongoose.connection;
+db.on('error', error => console.error('Connection to database unsuccessful', error));
+db.once('open', () => console.log(`Connection to database successful: ${MONGO_URI}`));
+
+const movieSchema = mongoose.Schema({
+  movieId: Number,
+  favorites: {
+    type: Number,
+    default: 0,
+  },
+  reviews: {
+    type: [{
+      message: String,
+      userId: Number,
+    }],
+    default: [],
+  },
+});
+
+const Movie = mongoose.model('Movie', movieSchema);
+
+module.exports.Movie = Movie;
